requirements.txt
asgiref==3.7.2
Django==3.2.10
django-ipware==6.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
python-ipware==2.0.1
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.9.0
